move module into pyspark_log_parser folder
